tvOS conditional compilation fixes
