+++ Sources/Honeycomb/MetricKitSubscriber.swift
@@ -1,3 +1,4 @@
+#if canImport(MXMetricManager)
 import Foundation
 import MetricKit
 import OpenTelemetryApi
@@ -19,55 +20,6 @@ class MetricKitSubscriber: NSObject, MXMetricManagerSubscriber {
         for payload in payloads {
             reportDiagnostics(payload: payload)
         }
-    }
-}
-
-// MARK: - AttributeValue helpers
-
-/// A protocol to make it easier to write generic functions for AttributeValues.
-protocol AttributeValueConvertable {
-    func attributeValue() -> AttributeValue
-}
-
-extension Int: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        AttributeValue.int(self)
-    }
-}
-extension Bool: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        AttributeValue.bool(self)
-    }
-}
-extension String: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        AttributeValue.string(self)
-    }
-}
-
-extension [String]: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        AttributeValue.array(AttributeArray(values: self.map { AttributeValue.string($0) }))
-    }
-}
-
-extension TimeInterval: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        // The OTel standard for time durations is seconds, which is also what TimeInterval is.
-        // https://opentelemetry.io/docs/specs/semconv/general/metrics/
-        AttributeValue.double(self)
-    }
-}
-extension Measurement: AttributeValueConvertable {
-    func attributeValue() -> AttributeValue {
-        // Convert to the "base unit", such as seconds or bytes.
-        let value =
-            if let unit = self.unit as? Dimension {
-                unit.converter.baseUnitValue(fromValue: self.value)
-            } else {
-                self.value
-            }
-        return AttributeValue.double(value)
     }
 }
 
@@ -464,3 +416,4 @@ public func reportDiagnostics(payload: MXDiagnosticPayload) {
         return attrs
     }
 }
+#endif
